+++ lib/oli/delivery/attempts/core.ex
@@ -416,11 +416,15 @@ defmodule Oli.Delivery.Attempts.Core do
       from(aa in ActivityAttempt,
         left_join: aa2 in ActivityAttempt,
         on:
-        aa2.resource_id == ^resource_id and aa.resource_attempt_id == aa2.resource_attempt_id and aa.id < aa2.id,
-        where: aa.resource_id == ^resource_id and aa.resource_attempt_id == ^resource_attempt_id and is_nil(aa2),
+          aa2.resource_id == ^resource_id and aa.resource_attempt_id == aa2.resource_attempt_id and
+            aa.id < aa2.id,
+        where:
+          aa.resource_id == ^resource_id and aa.resource_attempt_id == ^resource_attempt_id and
+            is_nil(aa2),
         select: aa
       )
-    ) |> Repo.preload(revision: [:activity_type])
+    )
+    |> Repo.preload(revision: [:activity_type])
   end
 
   @doc """