+++ lib/oli_web/controllers/page_delivery_controller.ex
@@ -435,12 +435,19 @@ defmodule OliWeb.PageDeliveryController do
     render(conn, "error.html")
   end
 
-  # This case handles :in_progress and :revised progress states
+  # This case handles :in_progress and :revised progress states, in addition to
+  # handling review mode
   defp render_page(%PageContext{} = context, conn, section_slug, user, _) do
     section = conn.assigns.section
 
     render_context = %Context{
-      user: user,
+      # Allow admin authors to review student work
+      user:
+        if is_nil(user) do
+          conn.assigns.current_author
+        else
+          user
+        end,
       section_slug: section_slug,
       mode:
         if context.review_mode do
@@ -526,8 +533,12 @@ defmodule OliWeb.PageDeliveryController do
         "attempt_guid" => attempt_guid
       }) do
     user = conn.assigns.current_user
+    author = conn.assigns.current_author
 
-    if Sections.is_enrolled?(user.id, section_slug) do
+    section = conn.assigns.section
+
+    if Oli.Accounts.is_admin?(author) or
+         PageLifecycle.can_access_attempt?(attempt_guid, user, section) do
       PageContext.create_for_review(section_slug, attempt_guid, user)
       |> render_page(conn, section_slug, user, false)
     else